Moved broke the user controller into two parts.  The admin related controllers are now separate.
Major concerns with the number of calls done on each screen.  This is not efficient and has to be broken up into several parts in order for this to scale.  Poor commenting, poorly constructed files, very little code reuse.

Overall a nightmare of choices in the construction of this administrative application.  I will have to revisit at another time.